requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.6
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
xlrd==2.0.1
xlwt==1.3.0
